Migrate databus custom-serializer to NServiceBus 10
- Updated all .csproj files to target .NET 10 and NServiceBus 10.0.0-alpha.1
- Added NServiceBus.ClaimCheck 2.0.0-alpha.1 package
- Migrated from DataBusProperty<T> to regular array properties
- Updated BsonDataBusSerializer to BsonClaimCheckSerializer implementing IClaimCheckSerializer
- Replaced UseDataBus with UseClaimCheck<FileShareClaimCheck, BsonClaimCheckSerializer>
- Added prerelease.txt marker file
- Removed InputLoopService pattern and moved logic to Program.cs
- Updated message handler to work with regular arrays instead of DataBusProperty<T>
- Added claim check conventions for properties starting with 'Large'

Fixes: #7289
