+++ samples/databus/custom-serializer/Core_10/Receiver/Program.cs
@@ -1,21 +1,23 @@
+using System;
 using Microsoft.Extensions.Hosting;
 using NServiceBus;
+using NServiceBus.ClaimCheck;
 using Shared;
-using System;
-using System.Threading.Tasks;
 
 Console.Title = "Receiver";
 var builder = Host.CreateApplicationBuilder(args);
 var endpointConfiguration = new EndpointConfiguration("Samples.DataBus.Receiver");
 
-#pragma warning disable CS0618 // Type or member is obsolete
 #region ConfigureReceiverCustomDataBusSerializer
 
-var dataBus = endpointConfiguration.UseDataBus<FileShareDataBus, BsonDataBusSerializer>();
-dataBus.BasePath(@"..\..\..\..\storage");
+var claimCheck = endpointConfiguration.UseClaimCheck<FileShareClaimCheck, BsonClaimCheckSerializer>();
+claimCheck.BasePath(@"..\..\..\..\storage");
 
 #endregion
-#pragma warning restore CS0618 // Type or member is obsolete
+
+#region ClaimCheckConventions
+endpointConfiguration.Conventions().DefiningClaimCheckPropertiesAs(prop => prop.Name.StartsWith("Large"));
+#endregion
 
 endpointConfiguration.UseSerialization<SystemJsonSerializer>();
 endpointConfiguration.UseTransport(new LearningTransport());
