+++ samples/databus/custom-serializer/Core_10/Sender/Program.cs
@@ -1,30 +1,78 @@
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using NServiceBus;
-using Sender;
-using Shared;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using NServiceBus;
+using NServiceBus.ClaimCheck;
+using Shared;
 
 Console.Title = "Sender";
 var builder = Host.CreateApplicationBuilder(args);
-builder.Services.AddHostedService<InputLoopService>();
 var endpointConfiguration = new EndpointConfiguration("Samples.DataBus.Sender");
 
-#pragma warning disable CS0618 // Type or member is obsolete
 #region ConfigureSenderCustomDataBusSerializer
 
-var dataBus = endpointConfiguration.UseDataBus<FileShareDataBus, BsonDataBusSerializer>();
-dataBus.BasePath(@"..\..\..\..\storage");
+var claimCheck = endpointConfiguration.UseClaimCheck<FileShareClaimCheck, BsonClaimCheckSerializer>();
+claimCheck.BasePath(@"..\..\..\..\storage");
 
 #endregion
-#pragma warning restore CS0618 // Type or member is obsolete
+
+#region ClaimCheckConventions
+endpointConfiguration.Conventions().DefiningClaimCheckPropertiesAs(prop => prop.Name.StartsWith("Large"));
+#endregion
 
 endpointConfiguration.UseSerialization<SystemJsonSerializer>();
 endpointConfiguration.UseTransport(new LearningTransport());
 builder.UseNServiceBus(endpointConfiguration);
 
-await builder.Build().RunAsync();
+var app = builder.Build();
+
+await app.StartAsync();
+
+var messageSession = app.Services.GetRequiredService<IMessageSession>();
+
+Console.WriteLine("Press Enter to send a large message with claim check");
+Console.WriteLine("Press any other key to exit");
+
+while (true)
+{
+    var key = Console.ReadKey(true);
+    Console.WriteLine();
+
+    if (key.Key == ConsoleKey.Enter)
+    {
+        await SendMessageLargePayload(messageSession);
+        continue;
+    }
+    break;
+}
+
+await app.StopAsync();
+
+static Task SendMessageLargePayload(IMessageSession messageSession)
+{
+    var measurements = GetMeasurements().ToArray();
+
+    var message = new MessageWithLargePayload
+    {
+        SomeProperty = "This message contains a large collection that will be sent on the claim check",
+        LargeData = measurements
+    };
+    Console.WriteLine(@"Message sent, the payload is stored in: ..\..\..\storage");
+    return messageSession.Send("Samples.DataBus.Receiver", message);
+}
+
+static IEnumerable<Measurement> GetMeasurements()
+{
+    for (var i = 0; i < 10000; i++)
+    {
+        yield return new Measurement
+        {
+            Timestamp = DateTimeOffset.UtcNow,
+            MeasurementName = $"Instrument {i}",
+            MeasurementValue = i * 10m
+        };
+    }
+}
