+++ samples/databus/custom-serializer/Core_10/Shared/MessageWithLargePayload.cs
@@ -6,9 +6,7 @@ public class MessageWithLargePayload :
     ICommand
 {
     public string SomeProperty { get; set; }
-#pragma warning disable CS0618 // Type or member is obsolete
-    public DataBusProperty<Measurement[]> LargeData { get; set; }
-#pragma warning restore CS0618 // Type or member is obsolete
+    public Measurement[] LargeData { get; set; }
 }
 
 [Serializable]
